+++ packages/api/src/codegen/factory.ts
@@ -3,7 +3,9 @@ import type Oas from 'oas';
 
 import TSGenerator from './languages/typescript/index.js';
 
-export type SupportedLanguages = 'js' | 'js-cjs' | 'js-esm' | 'ts';
+export enum SupportedLanguages {
+  JS = 'js',
+}
 
 export default function codegenFactory(
   language: SupportedLanguages,
@@ -12,18 +14,8 @@ export default function codegenFactory(
   identifier: string,
 ): CodeGenerator {
   switch (language) {
-    case 'js':
-      throw new TypeError('An export format of CommonJS or ECMAScript is required for JavaScript compilation.');
-
-    case 'js-cjs':
-    case 'js-esm':
-    case 'ts':
-      return new TSGenerator(spec, specPath, identifier, {
-        outputJS: ['js-cjs', 'js-esm'].includes(language),
-
-        // TS will always generate with ESM-like exports.
-        compilerTarget: language === 'js-cjs' ? 'cjs' : 'esm',
-      });
+    case SupportedLanguages.JS:
+      return new TSGenerator(spec, specPath, identifier);
 
     default:
       throw new TypeError(`Unsupported language supplied: ${language}`);