+++ packages/api/test/__fixtures__/sdk/response-title-quirks/src/index.ts
@@ -1,8 +1,8 @@
 import type * as types from './types';
-import type { ConfigOptions, FetchResponse } from '@readme/api-core';
+import type { ConfigOptions, FetchResponse, HTTPMethodRange } from '@readme/api-core';
 import Oas from 'oas';
 import APICore from '@readme/api-core';
-import definition from '@api/test-utils/definitions/simple.json';
+import definition from '@api/test-utils/definitions/response-title-quirks.json';
 
 class SDK {
   spec: Oas;
@@ -10,7 +10,7 @@ class SDK {
 
   constructor() {
     this.spec = Oas.init(definition);
-    this.core = new APICore(this.spec, 'simple-ts/1.0.0 (api/<<package version>>)');
+    this.core = new APICore(this.spec, 'response-title-quirks/1.0.0 (api/<<package version>>)');
   }
 
   /**
@@ -73,13 +73,8 @@ class SDK {
     this.core.setServer(url, variables);
   }
 
-  /**
-   * Multiple status values can be provided with comma separated strings
-   *
-   * @summary Finds Pets by status
-   */
-  findPetsByStatus(metadata: types.FindPetsByStatusMetadataParam): Promise<FetchResponse<200, types.FindPetsByStatusResponse200>> {
-    return this.core.fetch('/pet/findByStatus', 'get', metadata);
+  getAnything(metadata: types.GetAnythingMetadataParam): Promise<FetchResponse<HTTPMethodRange<200, 299>, types.GetAnythingResponse2XX>> {
+    return this.core.fetch('/anything', 'get', metadata);
   }
 }
 
@@ -87,5 +82,3 @@ const createSDK = (() => { return new SDK(); })()
 ;
 
 export default createSDK;
-
-export type { Category, FindPetsByStatusMetadataParam, FindPetsByStatusResponse200, Pet, Tag } from './types';