+++ packages/api/test/__fixtures__/sdk/simple-js-cjs/Category.js
@@ -1,4 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var Category = { "type": "object", "properties": { "id": { "type": "integer", "format": "int64", "minimum": -9223372036854776000, "maximum": 9223372036854776000 }, "name": { "type": "string" } }, "title": "Category", "x-readme-ref-name": "Category" };
-exports.default = Category;