+++ packages/api/test/__fixtures__/sdk/simple-js-cjs/schemas.d.ts
@@ -1,5 +1,0 @@
-import Category from './schemas/Category';
-import FindPetsByStatus from './schemas/FindPetsByStatus';
-import Pet from './schemas/Pet';
-import Tag from './schemas/Tag';
-export { Category, FindPetsByStatus, Pet, Tag };