+++ packages/api/test/__fixtures__/sdk/simple-js-esm/Category.js
@@ -1,2 +1,0 @@
-const Category = { "type": "object", "properties": { "id": { "type": "integer", "format": "int64", "minimum": -9223372036854776000, "maximum": 9223372036854776000 }, "name": { "type": "string" } }, "title": "Category", "x-readme-ref-name": "Category" };
-export default Category;