+++ packages/api/test/__fixtures__/sdk/simple-js-esm/Tag.js
@@ -1,2 +1,0 @@
-const Tag = { "type": "object", "properties": { "id": { "type": "integer", "format": "int64", "minimum": -9223372036854776000, "maximum": 9223372036854776000 }, "name": { "type": "string" } }, "title": "Tag", "x-readme-ref-name": "Tag" };
-export default Tag;