+++ packages/api/test/codegen/languages/typescript/index.test.ts
@@ -1,9 +1,8 @@
-import type { TSGeneratorOptions } from '../../../../src/codegen/languages/typescript/index.js';
-
 import fs from 'node:fs/promises';
 import path from 'node:path';
 
 import { loadSpec, responses as mockResponse } from '@api/test-utils';
+import execa from 'execa';
 import fetchMock from 'fetch-mock';
 import Oas from 'oas';
 import uniqueTempDir from 'unique-temp-dir';
@@ -13,13 +12,13 @@ import TSGenerator from '../../../../src/codegen/languages/typescript/index.js';
 import * as packageInfo from '../../../../src/packageInfo.js';
 import Storage from '../../../../src/storage.js';
 
-function assertSDKFixture(file: string, fixture: string, opts: TSGeneratorOptions = {}) {
+function assertSDKFixture(file: string, fixture: string) {
   return async () => {
     const oas = await loadSpec(require.resolve(file)).then(Oas.init);
     await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
 
-    const ts = new TSGenerator(oas, file, fixture, opts);
-    const actualFiles = await ts.compile();
+    const ts = new TSGenerator(oas, file, fixture);
+    const actualFiles = await ts.generate();
 
     // Determine if the generated code matches what we've got in our fixture.
     const dir = path.resolve(path.join(__dirname, '..', '..', '..', '__fixtures__', 'sdk', fixture));
@@ -28,11 +27,11 @@ function assertSDKFixture(file: string, fixture: string, opts: TSGeneratorOption
       .then(files => {
         files.sort();
 
-        // The `schemas` directory comes back to us from `fs.readdir` but because we're doing a
-        // recursive lookup on that we also get `schemas/<schemaName>`. We only care about the
-        // schema file itself, not the general directory, so we need to exclude this from our list
-        // of expected files.
-        return files.filter(f => f !== 'schemas');
+        // The `src/schemas` directory comes back to us from `fs.readdir` but because we're doing a
+        // recursive lookup on that we also get `src/schemas/<schemaName>`. We only care about the
+        // schema files themselves, not the general directory, so we need to exclude this from our
+        // list of expected files.
+        return files.filter(f => !['src', 'src/schemas'].includes(f));
       })
       .catch(() => {
         /**
@@ -54,26 +53,23 @@ function assertSDKFixture(file: string, fixture: string, opts: TSGeneratorOption
         // We have to wrap in our current package version into the `<<useragent>>` placeholder so
         // we don't need to worry about committing package versions into source control or trying
         // to mock out our `packageInfo` library, potentially causing sideeffects in other tests.
-        return new Promise((resolve, reject) => {
-          fs.readFile(path.join(dir, filename), 'utf8')
-            .then(expected => expected.replace('<<package version>>', packageInfo.PACKAGE_VERSION))
-            .then(expected => {
-              expect(actual).toBe(expected);
-              resolve(true);
-            })
-            .catch(reject);
-        });
+        return fs
+          .readFile(path.join(dir, filename), 'utf8')
+          .then(expected => expected.replace('<<package version>>', packageInfo.PACKAGE_VERSION))
+          .then(expected => {
+            expect(actual).toBe(expected);
+          });
       }),
     );
 
     // Make sure that we can load the SDK without any TS compilation errors.
-    const sdk = await import(`../../../__fixtures__/sdk/${fixture}`).then(r => r.default);
+    const sdk = await import(`${dir}/src/index.ts`).then(r => r.default);
     expect(sdk.constructor.name).toBe('SDK');
   };
 }
 
 describe('typescript', () => {
-  describe('#installer', () => {
+  describe('#install', () => {
     beforeEach(() => {
       Storage.setStorageDir(uniqueTempDir());
     });
@@ -82,7 +78,7 @@ describe('typescript', () => {
       Storage.reset();
     });
 
-    it('should install a `package.json` and the required packages', async () => {
+    it('should install the required packages', async () => {
       const logger = vi.fn();
 
       const file = require.resolve('@readme/oas-examples/3.0/json/petstore.json');
@@ -92,21 +88,16 @@ describe('typescript', () => {
       const storage = new Storage(file, 'petstore');
       await storage.load();
 
-      const ts = new TSGenerator(oas, './openapi.json', 'petstore', { compilerTarget: 'cjs' });
+      const ts = new TSGenerator(oas, '../openapi.json', 'petstore');
+      const source = await ts.generate();
+      await storage.saveSourceFiles(source);
+
       await ts.install(storage, { logger, dryRun: true });
 
-      const pkgJson = await fs
-        .readFile(path.join(storage.getIdentifierStorageDir(), 'package.json'), 'utf-8')
-        .then(JSON.parse);
-
-      expect(pkgJson).toStrictEqual({
-        name: '@api/petstore',
-        version: '1.0.0',
-        main: './index.ts',
-        types: './index.d.ts',
-      });
-
-      expect(logger).toHaveBeenCalledWith('npm install --save --dry-run api json-schema-to-ts oas');
+      const { stdout } = await execa('ls', ['-R1'], { cwd: storage.getIdentifierStorageDir() });
+      expect(stdout).toContain('index.ts');
+      expect(stdout).toContain('schemas');
+      expect(stdout).toContain('Pet.ts');
 
       /**
        * NPM has an incredibly difficult time trying to resolve this temp dir when installing
@@ -118,15 +109,12 @@ describe('typescript', () => {
        *
        * @fixme
        */
-      // expect(logger).to.be.calledWith(`npm install --save --dry-run ${storage.getIdentifierStorageDir()}`);
+      // expect(logger).toHaveBeenCalledWith(`npm install --save --dry-run ${storage.getIdentifierStorageDir()}`);
     }, 20000);
   });
 
-  describe('#generator', () => {
-    it(
-      'should generate typescript (by default)',
-      assertSDKFixture('@api/test-utils/definitions/simple.json', 'simple-ts'),
-    );
+  describe('#compile', () => {
+    it('should generate an sdk', assertSDKFixture('@api/test-utils/definitions/simple.json', 'simple'));
 
     it(
       'should be able to generate valid TS when a body is optional but metadata isnt',
@@ -155,30 +143,17 @@ describe('typescript', () => {
       assertSDKFixture('@api/test-utils/definitions/alby.json', 'alby'),
     );
 
-    describe('javascript generation', () => {
-      it(
-        'should generate a CommonJS library',
-        assertSDKFixture('@api/test-utils/definitions/simple.json', 'simple-js-cjs', { outputJS: true }),
-      );
-
-      it(
-        'should generate am ESM library',
-        assertSDKFixture('@api/test-utils/definitions/simple.json', 'simple-js-esm', {
-          outputJS: true,
-          compilerTarget: 'esm',
-        }),
-      );
-    });
-
     describe('integration', () => {
       afterEach(() => {
         fetchMock.restore();
       });
 
-      it('should be able to make an API request (TS)', async () => {
-        const sdk = await import('../../../__fixtures__/sdk/simple-ts/index.js').then(r => r.default);
+      it('should be able to make an API request', async () => {
+        // eslint-disable-next-line import/no-relative-packages
+        const sdk = await import('../../../__fixtures__/sdk/simple/src/index.js').then(r => r.default);
         fetchMock.get('http://petstore.swagger.io/v2/pet/findByStatus?status=available', mockResponse.searchParams);
 
+        // @ts-expect-error `findPetsByStatus` types are only present with a fully built out SDK dist which this fixture lacks.
         await sdk.findPetsByStatus({ status: ['available'] }).then(({ data, status, headers, res }) => {
           expect(data).toBe('/v2/pet/findByStatus?status=available');
           expect(status).toBe(200);
@@ -188,10 +163,12 @@ describe('typescript', () => {
       });
 
       it('should be able to make an API request with an `accept` header`', async () => {
-        const sdk = await import('../../../__fixtures__/sdk/simple-ts/index.js').then(r => r.default);
+        // eslint-disable-next-line import/no-relative-packages
+        const sdk = await import('../../../__fixtures__/sdk/simple/src/index.js').then(r => r.default);
         fetchMock.get('http://petstore.swagger.io/v2/pet/findByStatus?status=available', mockResponse.headers);
 
         await sdk
+          // @ts-expect-error `findPetsByStatus` types are only present with a fully built out SDK dist which this fixture lacks.
           .findPetsByStatus({ status: ['available'], accept: 'application/xml' })
           .then(({ data, status, headers, res }) => {
             expect(data).toHaveProperty('accept', 'application/xml');
@@ -199,37 +176,6 @@ describe('typescript', () => {
             expect(headers.constructor.name).toBe('Headers');
             expect(res.constructor.name).toBe('Response');
           });
-      });
-
-      /**
-       * This test is impossible to run now because its a `.js` file that's loading ESM code. The
-       * CJS SDK we're generating here should have a `.cjs` extension but we're going to overhaul
-       * this entire codegen process with `tsup` so this test is being skipped for now.
-       *
-       * @see {@link https://github.com/readmeio/api/pull/734}
-       */
-      // it.skip('should be able to make an API request (JS + CommonJS)', async () => {
-      //   const sdk = await import('../../../__fixtures__/sdk/simple-js-cjs/index.js').then(r => r.default);
-      //   fetchMock.get('http://petstore.swagger.io/v2/pet/findByStatus?status=available', mockResponse.searchParams);
-
-      //   await sdk.findPetsByStatus({ status: ['available'] }).then(({ data, status, headers, res }) => {
-      //     expect(data).toBe('/v2/pet/findByStatus?status=available');
-      //     expect(status).toBe(200);
-      //     expect(headers.constructor.name).toBe('Headers');
-      //     expect(res.constructor.name).toBe('Response');
-      //   });
-      // });
-
-      it('should be able to make an API request (JS + ESM)', async () => {
-        const sdk = await import('../../../__fixtures__/sdk/simple-js-esm/index.js').then(r => r.default);
-        fetchMock.get('http://petstore.swagger.io/v2/pet/findByStatus?status=available', mockResponse.searchParams);
-
-        await sdk.findPetsByStatus({ status: ['available'] }).then(({ data, status, headers, res }) => {
-          expect(data).toBe('/v2/pet/findByStatus?status=available');
-          expect(status).toBe(200);
-          expect(headers.constructor.name).toBe('Headers');
-          expect(res.constructor.name).toBe('Response');
-        });
       });
     });
 
@@ -245,7 +191,7 @@ describe('typescript', () => {
         });
 
         const ts = new TSGenerator(oas, 'no-paths', './no-paths.json');
-        await expect(ts.compile()).rejects.toThrow(
+        await expect(ts.generate()).rejects.toThrow(
           'Sorry, this OpenAPI definition does not have any operation paths to generate an SDK for.',
         );
       });